refactor(build): extract phase constants from v6_build (#2747 step 1)
The Monitor API (state_build/state_helpers/state_issues/state_coverage/main)
imported PHASES/PHASE_LABELS from scripts/build/v6_build.py — pulling the
entire 12,299-line OBSOLETE build entrypoint into the API process just to read
a phase list. This was the load-bearing coupling flagged in #2747 (A) that
makes v6_build un-removable.

Move PHASES/PHASE_LABELS (pure data) to a new dependency-free module
scripts/build/phase_constants.py. v6_build re-imports them (ALL_PHASES as
_ALL_PHASES, PHASES, PHASE_LABELS) so its internal resume/state helpers and
its full test suite are unaffected — object identity is preserved
(PHASES is v6_build.PHASES is phase_constants.PHASES). Repointed the 5 live
API modules + 2 API tests to the new module.

This is step 1 of the #2747 migration: the Monitor API no longer depends on
v6_build. Remaining before v6_build/v5 can be removed: extract the 2 prompt
helpers (_strip_prompt_control_tags/_format_prompt_literal_block, used by the
V7 plan_patch phase) and _post_process_content (post-processor migration), then
migrate the research-preseed pipeline off build_module_v5.

Verified: ruff clean; import identity smoke; 343 tests pass across
tests/test_api_*, tests/api/, test_v6_build_events, test_manifest_api,
test_orient_api, test_api_state.

X-Agent: claude-inline/extract-phase-constants
